Final version of AddProblem tests
